resultset: fix agent escape string
If agent name had `'` in it, it was not escaped properly. This replaces `'` with `\'` so the string doesn't escape early.

For example `(agent = 'Baldur's Gate Dark Alliance')` broke the query.

Testing: Make a desktop file in `~/.local/applications` with app name that has `'` in it and run plasmashell with ` QT_LOGGING_RULES="*.debug=true;qt.*=false;org.kde.plasma.pulseaudio=false" plasmashell --replace`