resultmodel: guard model in threaded resource cleanup
this previously was subject to a race condition between ResultModel
destruction and the file access finishing. When the file access took
long enough the Model would have been deleted leading to a crash.

Instead weakly point to the model via the QObject such that we can then
check if still exists by the time we want to call into it

BUG: 458121
BUG: 466995
Fixes PLASMA-WORKSPACE-YZ